[FEATURE] Add display_layout submodules.
